+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 0.13.0
 ------
 
+- feat: `Iso` support `ListIsos`, `AttachIso`, and `DetachIso`
 - change: `Paginate` to accept `Listable`
 - change: `ListCommand` is also `Listable`
 - change: `client.Get` doesn't modify the given resource, returns a new one